refactor: delay 시간 상수로 뺴기
