+++ docs/README.md
@@ -29,10 +29,6 @@
 - [x] 정상적으로 게임의 턴이 다 동작된 후에는 결과를 보여주고, 2초 후에 축하의 alert 메세지를 띄운다.
 - [ ] 위 기능들이 정상적으로 동작하는지 Cypress를 이용해 테스트한다.
 
-### refactoring
-- [x] 파일 분리
-- [x] index.html에 `<div id="app"></div>`만 남기기.
-
 ### 버그 수정
 - [ ] input 입력한 뒤 비활성화 하기.
