--- src/js/modules/@share/message.ts
@@ -1,9 +1,0 @@
-export const ERROR_MESSAGE = {
-  INVALID_CAR_NAME_INPUT: '유효하지 않은 자동차 이름입니다. 다시 입력 해주세요.',
-  INVALID_INPUT_PROCEDURE: '자동차 이름 먼저 입력해주세요!',
-  INVALID_COUNT_INPUT: '유효하지 않은 횟수 입력입니다. 다시 입력 해주세요.'
-};
-
-export const MESSAGE = {
-    CELEBRATE_WINNER: '🏆 축하합니다 ㅎㅎ 최종 우승자: EAST, WEST, SOUTH, NORTH 🏆'
-}
